各テーブルのデータ型を変更

+++ database/migrations/2025_02_22_124420_create_filnal_results_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                     ->constrained('applies', 'apply_id')
                     ->onUpdate('cascade')
                     ->onDelete('cascade');
-            $table->integer('status');
+            $table->unsignedTinyInteger('status');
             $table->text('memo')->nullable();
             $table->dateTime('created_at')->useCurrent()->nullable();
             $table->dateTime('updated_at')->useCurrent()->nullable();
